remix and move drag and drop test
